Initial PVI framework // WIP
